Alteração do tipo do interest_rate em Loan.java e criação de LoanService s.java e modificação em OffersServices.java

+++ src/main/java/com/loans/loans/Model/Loan.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class Loan {
     private String type;
-    private BigDecimal interest_rate;
+    private int interest_rate;
 }
